apps/PMC跟仓管/配色库存管理/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
pandas==2.2.0
openpyxl==3.1.2
pytest==8.0.0
requests>=2.31
Pillow>=10
gunicorn>=21
